goit-cs-hw-03/task_02/requirements.txt
pymongo==4.3.3
python-dotenv==0.19.2
certifi==2024.8.30